integration_tests/test-requirements-for-tox.txt
codetiming
kombu
openapi-spec-validator<0.6.0  # dependency conflict on requests version (>2.31.0) with wazo clients (=2.25.1)
pyhamcrest
pytest
python-ldap
pytest-playwright

# for database tests
https://github.com/wazo-platform/wazo-lib-rest-client/archive/master.zip
https://github.com/wazo-platform/wazo-auth-client/archive/master.zip
https://github.com/wazo-platform/wazo-confd-client/archive/master.zip
https://github.com/wazo-platform/wazo-test-helpers/archive/master.zip
https://github.com/wazo-platform/xivo-lib-python/archive/master.zip
https://github.com/wazo-platform/wazo-bus/archive/master.zip
sqlalchemy<2.0
jinja2==2.11.3
psycopg2-binary
stevedore
unidecode
marshmallow
python-consul==1.1.0
